+++ src/anvil/lib/format_versions.py
@@ -1,7 +1,7 @@
 MANIFEST_BUILD: str = "1.20.80"  # The build version of the manifest.
-BLOCK_SERVER_VERSION: str = "1.21.70"  # The version of the block server.
-ENTITY_SERVER_VERSION: str = "1.21.70"  # The version of the entity server.
-ENTITY_CLIENT_VERSION: str = "1.21.70"  # The version of the entity client.
+BLOCK_SERVER_VERSION: str = "1.21.80"  # The version of the block server.
+ENTITY_SERVER_VERSION: str = "1.21.80"  # The version of the entity server.
+ENTITY_CLIENT_VERSION: str = "1.21.80"  # The version of the entity client.
 BP_ANIMATION_VERSION: str = "1.10.0"  # The version of the behavior pack animation.
 RP_ANIMATION_VERSION: str = "1.10.0"  # The version of the resource pack animation.
 ANIMATION_CONTROLLERS_VERSION: str = "1.10.0"  # The version of the animation controllers.
@@ -16,10 +16,10 @@ ITEM_SERVER_VERSION: str = "1.21.70"  # The version of the item server.
 GLOBAL_LIGHTING = "1.0.0"
 CAMERA_PRESET_VERSION = "1.21.0"  # The version of the camera presets.
 BLOCK_JSON_FORMAT_VERSION = "1.1.0"  # The version of the block json format.
-RECIPE_JSON_FORMAT_VERSION = "1.21.70"  # The version of the recipe json format.
-CRAFTING_ITEMS_CATALOG = "1.21.60" # The version of the crafting items catalog.
+RECIPE_JSON_FORMAT_VERSION = "1.21.80"  # The version of the recipe json format.
+CRAFTING_ITEMS_CATALOG = "1.21.60"  # The version of the crafting items catalog.
 
-MODULE_MINECRAFT_SERVER: str = "1.17.0"  # The version of the Minecraft server module.
+MODULE_MINECRAFT_SERVER: str = "1.19.0"  # The version of the Minecraft server module.
 MODULE_MINECRAFT_SERVER_UI: str = "1.3.0"  # The version of the Minecraft UI module.
 MODULE_MINECRAFT_SERVER_EDITOR: str = "0.1.0"  # The version of the Minecraft UI module.
 MODULE_MINECRAFT_SERVER_GAMETEST: str = "1.0.0"  # The version of the Minecraft UI module.